backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
python-dotenv==1.0.1
gunicorn==20.1.0
Flask-RESTful==0.3.9
numpy==1.26.4
Pillow==10.2.0
pdf2image==1.16.3
openpyxl==3.1.2
pandas==2.2.1